# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: QuickProp Office API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties:
    get:
      operationId: listProperties
      tags: [properties]
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: pipelineStage, in: query, schema: { type: string } }
        - { name: listingType, in: query, schema: { type: string } }
        - { name: propertyType, in: query, schema: { type: string } }
        - { name: suburb, in: query, schema: { type: string } }
        - { name: agentId, in: query, schema: { type: integer } }
        - { name: q, in: query, schema: { type: string } }
        - { name: minPrice, in: query, schema: { type: number } }
        - { name: maxPrice, in: query, schema: { type: number } }
      responses:
        "200":
          description: Properties
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Property" }
    post:
      operationId: createProperty
      tags: [properties]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PropertyInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Property" }
  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Property
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Property" }
        "404": { description: Not found }
    patch:
      operationId: updateProperty
      tags: [properties]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PropertyUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Property" }
    delete:
      operationId: deleteProperty
      tags: [properties]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
  /properties/{id}/price-history:
    get:
      operationId: getPropertyPriceHistory
      tags: [properties]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Price history
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PriceHistoryEntry" }
  /properties/{id}/activity:
    get:
      operationId: getPropertyActivity
      tags: [properties]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Activity timeline
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ActivityEntry" }
  /properties/{id}/share:
    post:
      operationId: shareProperty
      tags: [marketing]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ShareInput" }
      responses:
        "200":
          description: Share link
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShareResult" }
  /properties/{id}/brochure:
    post:
      operationId: generateBrochure
      tags: [marketing]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Brochure generated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BrochureResult" }
...
components:
  schemas:
...
    Property:
      type: object
      required: [id, reference, title, status, pipelineStage, listingType, propertyType, price, currency, suburb, city]
      properties:
        id: { type: integer }
        reference: { type: string }
        title: { type: string }
        description: { type: ["string", "null"] }
        propertyType: { type: string, description: "house | apartment | townhouse | stand | commercial | industrial | farm" }
        listingType: { type: string, description: "sale | rent" }
        status: { type: string, description: "draft | internal_only | public | under_offer | sold | rented | archived | withdrawn | coming_soon | private_listing" }
        pipelineStage: { type: string, description: "draft | awaiting_photos | ready | published | under_offer | sold | archived" }
        price: { type: number }
        currency: { type: string }
        suburb: { type: string }
        city: { type: string }
        address: { type: ["string", "null"] }
        bedrooms: { type: ["integer", "null"] }
        bathrooms: { type: ["integer", "null"] }
        parking: { type: ["integer", "null"] }
        landSize: { type: ["number", "null"], description: "square metres" }
        buildingSize: { type: ["number", "null"] }
        features:
          type: array
          items: { type: string }
        photos:
          type: array
          items: { type: string }
        coverImage: { type: ["string", "null"] }
        agentId: { type: ["integer", "null"] }
        branchId: { type: ["integer", "null"] }
        sellerId: { type: ["integer", "null"] }
        mandateType: { type: ["string", "null"], description: "open | sole | exclusive" }
        mandateStart: { type: ["string", "null"] }
        mandateExpiry: { type: ["string", "null"] }
        commissionPercent: { type: ["number", "null"] }
        privateNotes: { type: ["string", "null"] }
        views: { type: integer }
        enquiries: { type: integer }
        shares: { type: integer }
        hasBrochure: { type: boolean }
        publishedAt: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    PropertyInput:
      type: object
      required: [title, propertyType, listingType, price, suburb, city]
      properties:
        title: { type: string }
        description: { type: string }
        propertyType: { type: string }
        listingType: { type: string }
        status: { type: string }
        pipelineStage: { type: string }
        price: { type: number }
        currency: { type: string }
        suburb: { type: string }
        city: { type: string }
        address: { type: string }
        bedrooms: { type: integer }
        bathrooms: { type: integer }
        parking: { type: integer }
        landSize: { type: number }
        buildingSize: { type: number }
        features:
          type: array
          items: { type: string }
        photos:
          type: array
          items: { type: string }
        coverImage: { type: string }
        agentId: { type: integer }
        branchId: { type: integer }
        sellerId: { type: integer }
        mandateType: { type: string }
        mandateStart: { type: string }
        mandateExpiry: { type: string }
        commissionPercent: { type: number }
        privateNotes: { type: string }
    PropertyUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        propertyType: { type: string }
        listingType: { type: string }
        status: { type: string }
        pipelineStage: { type: string }
        price: { type: number }
        currency: { type: string }
        suburb: { type: string }
        city: { type: string }
        address: { type: string }
        bedrooms: { type: integer }
        bathrooms: { type: integer }
        parking: { type: integer }
        landSize: { type: number }
        buildingSize: { type: number }
        features:
          type: array
          items: { type: string }
        photos:
          type: array
          items: { type: string }
        coverImage: { type: string }
        agentId: { type: integer }
        branchId: { type: integer }
        sellerId: { type: integer }
        mandateType: { type: string }
        mandateStart: { type: string }
        mandateExpiry: { type: string }
        commissionPercent: { type: number }
        privateNotes: { type: string }
    PriceHistoryEntry:
      type: object
      required: [id, propertyId, price, changedAt]
      properties:
        id: { type: integer }
        propertyId: { type: integer }
        price: { type: number }
        previousPrice: { type: ["number", "null"] }
        changedAt: { type: string }
        changedBy: { type: ["string", "null"] }
    ActivityEntry:
      type: object
      required: [id, type, message, createdAt]
      properties:
        id: { type: integer }
        type: { type: string, description: "created | edited | status_change | price_change | published | photo_upload | viewing | enquiry | share | brochure | login" }
        message: { type: string }
        entityType: { type: ["string", "null"] }
        entityId: { type: ["integer", "null"] }
        userName: { type: ["string", "null"] }
        createdAt: { type: string }
...
    ShareInput:
      type: object
      required: [channel]
      properties:
        channel: { type: string, description: "link | whatsapp | facebook | email | sms" }
    ShareResult:
      type: object
      required: [url]
      properties:
        url: { type: string }
        qrData: { type: ["string", "null"] }
    BrochureResult:
      type: object
      required: [propertyId, generatedAt]
      properties:
        propertyId: { type: integer }
        url: { type: ["string", "null"] }
        generatedAt: { type: string }
...