# lib/api-zod/src/generated/types/brochureResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * QuickProp Office API
 * OpenAPI spec version: 0.1.0
 */

export interface BrochureResult {
  propertyId: number;
  /** @nullable */
  url?: string | null;
  generatedAt: string;
}