# lib/api-zod/src/generated/types/propertyUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * QuickProp Office API
 * OpenAPI spec version: 0.1.0
 */

export interface PropertyUpdate {
  title?: string;
  description?: string;
  propertyType?: string;
  listingType?: string;
  status?: string;
  pipelineStage?: string;
  price?: number;
  currency?: string;
  suburb?: string;
  city?: string;
  address?: string;
  bedrooms?: number;
  bathrooms?: number;
  parking?: number;
  landSize?: number;
  buildingSize?: number;
  features?: string[];
  photos?: string[];
  coverImage?: string;
  agentId?: number;
  branchId?: number;
  sellerId?: number;
  mandateType?: string;
  mandateStart?: string;
  mandateExpiry?: string;
  commissionPercent?: number;
  privateNotes?: string;
}